requirements.txt
asgiref==3.9.2
Django==5.2.7
djangorestframework==3.16.1
#gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1    # load .env file
psycopg
PyJWT